requirements.txt
flask
transformers
torch
mlflow
datasets
azureml-sdk
azure-identity
azure-storage-blob
azureml-core

